requirements.txt
openai==0.28.1
telethon==1.31.0
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
duckduckgo_search==3.9.3
unidecode==1.3.7
tiktoken==0.5.1
pytz==2023.3.post1
bardapi==0.1.29
EdgeGPT==0.12.1
coloredlogs==15.0.1
log2d==0.0.18
google-generativeai==0.7.2
Pillow==10.1.0
python-telegram-bot==20.7
